backend/requirements.txt
fastapi
uvicorn
pyautogen
pydantic
python-dotenv
